requirements.txt
wagtail>=1.13,<1.14
wagtailemoji==1.0.0
wagtailfontawesome==1.0.6
wagtailerrorpages==1.1.5
wagtailgridder==0.9.18
wagtailcodeblock==0.2.4
wagtailcontentstream==0.2.3
django-modelcluster>=3.1
git+https://github.com/FlipperPA/wagtailtinymce.git
django>=1.11,<1.12
inflection==0.3.1
psycopg2==2.7.3.2
django-taggit==0.22.1
Markdown==2.6.8
django-tinymce==2.6.0
django-braces==1.4.0
django-wysiwyg-redactor==0.5.1
django-photologue==3.7
django-haystack==2.6.1
pysolr==3.6.0

click==6.4
cov-core==1.15.0
coverage==3.7.1
django-debug-toolbar==1.8
django-template-debug==0.3.5
django-test-plus
invoke<0.13.0
ipython==5.0.0
mock==1.0.1
pytest==2.6.4
pytest-cov==1.8.1
pytest-django==2.9.1
